Add Context to Patient History Fix Back-End for Context : small changes Repo Service Controller Add validation for Note/History
Fix Front-End

+++ src/PatientHistory/Infrastructure/Services/NoteService.cs
@@ -3,7 +3,10 @@ using PatientHistory.Data.DTO;
 using PatientHistory.Infrastructure.Repositories;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using FluentValidation;
 using Mapster;
+using PatientHistory.Internal;
+using Serilog;
 
 namespace PatientHistory.Infrastructure.Services
 {
@@ -11,11 +14,13 @@ namespace PatientHistory.Infrastructure.Services
     {
         private readonly INoteRepository _noteRepository;
         private readonly IPatientService _patientService;
+        private readonly NoteValidator _noteValidator;
 
         public NoteService(INoteRepository noteRepository, IPatientService patientService)
         {
             _noteRepository = noteRepository;
             _patientService = patientService;
+            _noteValidator = new NoteValidator();
         }
 
         public List<NoteDto> Get()
@@ -49,20 +54,39 @@ namespace PatientHistory.Infrastructure.Services
             return notesByPatient;
         }
 
-        public async Task Create(Note note)
+        public async Task Create(NoteDto noteDto)
         {
-            bool patient = await _patientService.Get(note.PatientId);
+            var validationResult = await _noteValidator.ValidateAsync(noteDto);
+
+            if (!validationResult.IsValid)
+            {
+                throw new ValidationException(validationResult.ToString(), validationResult.Errors);
+            }
+
+            bool patient = await _patientService.Get(noteDto.PatientId);
             if (!patient)
             {
-                throw new KeyNotFoundException($"{note.PatientId}");
+                throw new KeyNotFoundException($"{noteDto.PatientId}");
             }
+
+            var note = noteDto.Adapt<Note>();
             _noteRepository.Create(note);
         }
 
-        public void Update(string id, Note note) =>
+        public void Update(string id, NoteDto noteDto)
+        {
+            var note = noteDto.Adapt<Note>();
             _noteRepository.Update(id, note);
+        }
 
-        public void Remove(string id) =>
+        public void Remove(string id)
+        {
+            var noteDto = Get(id);
+            if (noteDto is null)
+            {
+                throw new KeyNotFoundException($"{id}");
+            }
             _noteRepository.Remove(id);
+        }
     }
 }
